feat: Add movie creation API route
The code changes add a new API route for creating movies. The route handles a POST request and expects parameters such as title, duration, dateRelease, synopsis, note, poster, and directorId. Upon successful creation, the route returns the created movie as a JSON response. In case of an error, it logs the error and returns an "Internal Server Error" response with a status code of 500.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and providing a concise description of the change.