feat: Add PrismaClient instance to db module
The code changes add a PrismaClient instance to the db module. This allows the module to either use the global prisma instance or create a new PrismaClient instance. The purpose of this change is to ensure that the PrismaClient instance is available for database operations throughout the application. This commit follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and providing a concise description of the change.